added login to support directory in rspec
